+++ tests/OSPSuite.SimModel.Tests/src/ExplicitFormulaSpecs.cpp
@@ -383,7 +383,7 @@ namespace UnitTests
 
 				vector<double> switchTimePoints=f->SwitchTimePoints();
 
-				BDDExtensions::ShouldBeEqualTo(switchTimePoints.size(), (unsigned int)7);
+				BDDExtensions::ShouldBeEqualTo(switchTimePoints.size(), (size_t)7);
 
 				DoubleQueue dq;
